refactor: Comment functionality and improve security
-댓글 수정 및 삭제 시 사용자 권한 검사 로직을 추가
-@RequestHeader를 사용해 JWT 필터에서 설정한 userId를 직접 사용하도록 변경
-예외 처리를 강화하여 오류 상황을 더 명확하게 처리

+++ src/main/java/com/newsfeed/fakebook/service/CommentService.java
@@ -44,16 +44,24 @@ public class CommentService {
                 .collect(Collectors.toList());
     }
 
-    public void updateComment(Long commentId, CommentRequestDto requestDto) {
+    public void updateComment(Long userId, Long commentId, CommentRequestDto requestDto) throws IllegalAccessException {
         Comment comment = commentRepository.findById(commentId)
                 .orElseThrow(() -> new IllegalArgumentException("댓글을 찾을 수 없습니다."));
+
+        if (!comment.getUser().getUserId().equals(userId)) {
+            throw new IllegalAccessException("댓글을 수정할 권한이 없습니다.");
+        }
 
         comment.updateContent(requestDto.getContent());
     }
 
-    public void deleteComment(Long commentId) {
+    public void deleteComment(Long userId, Long commentId) throws IllegalAccessException {
         Comment comment = commentRepository.findById(commentId)
                 .orElseThrow(() -> new IllegalArgumentException("댓글을 찾을 수 없습니다."));
+
+        if (!comment.getUser().getUserId().equals(userId)) {
+            throw new IllegalAccessException("댓글을 삭제할 권한이 없습니다");
+        }
 
         commentRepository.delete(comment);
     }
